requirements.txt
flask~=3.0.3
python-dotenv~=0.21.0
openai~=2.21.0
pyautogui~=0.9.54
requests~=2.32.3